+++ instapy/instapy.py
@@ -14,6 +14,7 @@ from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 from selenium.webdriver import DesiredCapabilities
 from selenium.webdriver.common.proxy import Proxy, ProxyType
+from selenium.webdriver.firefox.options import Options as Firefox_Options
 from pyvirtualdisplay import Display
 import logging
 from contextlib import contextmanager
@@ -273,6 +274,10 @@ class InstaPy:
             return self
 
         if self.use_firefox:
+            firefox_options = Firefox_Options()
+            if self.headless_browser:
+                firefox_options.add_argument('-headless')
+
             if self.browser_profile_path is not None:
                 firefox_profile = webdriver.FirefoxProfile(
                     self.browser_profile_path)
@@ -295,7 +300,8 @@ class InstaPy:
                 firefox_profile.set_preference('network.proxy.ssl_port',
                                                self.proxy_port)
 
-            self.browser = webdriver.Firefox(firefox_profile=firefox_profile)
+            self.browser = webdriver.Firefox(firefox_profile=firefox_profile,
+                                             options=firefox_options)
 
         else:
             chromedriver_location = Settings.chromedriver_location
@@ -369,7 +375,7 @@ class InstaPy:
 
         return self
 
-      
+
 
     def set_selenium_remote_session(self, selenium_url='', selenium_driver=None):
         """
@@ -1792,9 +1798,11 @@ class InstaPy:
             try:
                 links = get_links_for_username(
                     self.browser,
+                    self.username,
                     username,
                     amount,
                     self.logger,
+                    self.logfolder,
                     randomize,
                     media)
 
@@ -1889,7 +1897,7 @@ class InstaPy:
                                     user_name not in self.dont_include and
                                     checked_img and
                                     commenting):
-                                    
+
                                 if self.delimit_commenting:
                                     (self.commenting_approved,
                                      disapproval_reason) = verify_commenting(
@@ -2030,9 +2038,11 @@ class InstaPy:
 
             try:
                 links = get_links_for_username(self.browser,
+                                               self.username,
                                                username,
                                                amount,
                                                self.logger,
+                                               self.logfolder,
                                                randomize,
                                                media)
             except NoSuchElementException:
@@ -2281,9 +2291,11 @@ class InstaPy:
 
             try:
                 links = get_links_for_username(self.browser,
+                                               self.username,
                                                username,
                                                amount,
                                                self.logger,
+                                               self.logfolder,
                                                randomize,
                                                media,
                                                taggedImages=True)
@@ -2354,14 +2366,7 @@ class InstaPy:
 
                                 if self.use_clarifai and commenting:
                                     try:
-                                        checked_img, temp_comments = (
-                                            check_image(self.browser,
-                                                        self.clarifai_api_key,
-                                                        self.clarifai_img_tags,
-                                                        self.clarifai_img_tags_skip,
-                                                        self.logger,
-                                                        self.clarifai_full_match))
-
+                                        checked_img, temp_comments, clarifai_tags = self.query_clarifai()
                                     except Exception as err:
                                         self.logger.error(
                                             'Image check error: {}'.format(err))