+++ instapy/util.py
@@ -2,7 +2,6 @@
 import random
 import time
 import datetime
-import random
 import re
 import signal
 import os
@@ -129,9 +128,19 @@ def validate_username(browser,
         return False, \
                "---> {} is in ignore_users list  ~skipping user\n".format(username)
 
-    if username in blacklist:
-        return False, \
-               "---> {} is in blacklist  ~skipping user\n".format(username)
+    logfolder = logfolder = '{0}{1}{2}{1}'.format(
+            Settings.log_location, os.path.sep, own_username)
+
+    blacklist_file = "{}blacklist.csv".format(logfolder)
+    blacklist_file_exists = os.path.isfile(blacklist_file)
+    if blacklist_file_exists:
+        with open("{}blacklist.csv".format(logfolder), 'rt') as f:
+            reader = csv.reader(f, delimiter=',')
+            for row in reader:
+                for field in row:
+                    if field == username:
+                        logger.info('Username in BlackList: {} '.format(username))
+                        return False, "---> {} is in blacklist  ~skipping user\n".format(username)
 
     """Checks the potential of target user by relationship status in order to delimit actions within the desired boundary"""
     if potency_ratio or delimit_by_numbers and (max_followers or max_following or min_followers or min_following):
@@ -1470,6 +1479,3 @@ def get_action_delay(action):
         return defaults[action]
 
     return custom_delay
-
-
-