+++ quickstart_templates/Basic follow-unfollow activity.py
@@ -15,8 +15,6 @@ NOTES:
 from instapy import InstaPy
 from instapy.util import smart_run
 
-
-
 # login credentials
 insta_username = ''
 insta_password = ''
@@ -36,11 +34,11 @@ with smart_run(session):
                                       max_followers=4590,
                                       min_followers=45,
                                       min_following=77)
-    
+
     session.set_dont_include(["friend1", "friend2", "friend3"])
     session.set_dont_like(["pizza", "#store"])
-    
-    
+
+
     # activities
 
     """ Massive Follow of users followers (I suggest to follow not less than 3500/4000 users for better results)...
@@ -50,7 +48,7 @@ with smart_run(session):
     """ First step of Unfollow action - Unfollow not follower users...
     """
     session.unfollow_users(amount=500, InstapyFollowed=(True, "nonfollowers"), style="FIFO", unfollow_after=12*60*60, sleep_delay=601)
-        
+
     """ Second step of Massive Follow...
     """
     session.follow_user_followers(['user1','user2','user3'], amount=800, randomize=False, interact=False)
@@ -61,7 +59,4 @@ with smart_run(session):
 
     """ Clean all followed user - Unfollow all users followed by InstaPy...
     """
-    session.unfollow_users(amount=500, InstapyFollowed=(True, "all"), style="FIFO", unfollow_after=24*60*60, sleep_delay=601) 
-
-
-
+    session.unfollow_users(amount=500, InstapyFollowed=(True, "all"), style="FIFO", unfollow_after=24*60*60, sleep_delay=601)