+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The **goal** of this file is explaining to the users of our project the notable 
 _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)_.
 
 
+## [Unreleased] - 2018-11-07
+### Changed
+- Maintain names: 'person' for target user and 'username' for our running user
+- Verify private users in get_links_for_username
+- Changed behaviour of validate_username to check if a user is included in the blacklist. If yes will skip it and log why.
+
+
 ## [Unreleased] - 2018-11-01
 ### Added
 - Interact with tagged images of users, and validation of a user to be optional